refactor(FE-11): added ViewModel hook for StopMarker
