+++ src/features/stops/model/stop-marker/types.ts
@@ -1,7 +1,10 @@
-// export interface IStopMarkerViewModelProps {
-    
-// }
+import { Holiday } from "entities/holidays"
+import { Stop } from "entities/stop"
+import { Weather } from "entities/weather"
 
-// export interface IStopMarkerViewModel {
-
-// }
+export interface StopMarkerViewModel {
+    weather: Weather | null
+    holiday: Holiday | null
+    isLoading: boolean
+    updateStopInfo: (stop: Stop) => Promise<void>
+}
